requirements.txt
python-magic

